Fixed css, added loading handler

+++ phonebook-angular/src/app/contact-list/contact-list.component.ts
@@ -10,6 +10,7 @@ import { ArchitectService } from '../services/architect.service';
 })
 export class ContactListComponent implements OnInit {
   contacts: Contact[] = [];
+  deleteInProgress: string[] = [];
   constructor(private _architectService: ArchitectService) {}
 
   ngOnInit(): void {
@@ -18,14 +19,24 @@ export class ContactListComponent implements OnInit {
       .then((contacts) => (this.contacts = contacts));
   }
 
+  isContactBeingDeleted(id: string) {
+    return this.deleteInProgress.includes(id);
+  }
+
   async deleteContact(e: Event, id: string) {
     e.preventDefault();
     e.stopPropagation();
+    if (this.isContactBeingDeleted(id)) return;
+    this.deleteInProgress = [...this.deleteInProgress, id];
     try {
       await this._architectService.deleteContact(id);
       this.contacts = this.contacts.filter((contact) => contact.id !== id);
     } catch (e) {
       console.log(e);
+    } finally {
+      this.deleteInProgress = this.deleteInProgress.filter(
+        (deleteId) => deleteId !== id,
+      );
     }
   }
 }
